[core] search process impl to gatehr residuals. encoder.h: change the outline; predictor.c: avoid redundant cfl_alpha calc, the flag may be set for residual calc; search.c: change logic of mode info comparison; micand helper funcs: c1enc_block_has_xxx_cand; make gather_rdstat recursive; impl block level gather funcs(pred_type, residual); init pred_opt with {0} to avd uninit new fields; fix uninit matrices in search_inter_b; dev c1enc_search_sb logic (1/2); types: deprecate mvd field for inter mi; add ref frame getter helper func;

+++ src/encode/encoder.c
@@ -401,7 +401,7 @@ static c1enc_ref_t *c1enc__ref_at_fromref(c1enc_ref_t *ref, uint8_t y, uint8_t x
     }
 }
 c1enc_ref_t *c1enc_ref_at(c1enc_ctx_t *ctx, uint16_t sb_y, uint16_t sb_x, uint8_t y, uint8_t x, uint8_t ref_id) {
-    const int idx = ctx->avail_ref_cnt - ref_id;
+    const int idx = (int)ctx->avail_ref_cnt - ref_id;
     assert_fatal(idx >= 0 && idx < ctx->avail_ref_cnt);
     const uint16_t h = C1_ROUND_UP(ctx->ref_frames[idx].h, 64);
     const uint16_t w = C1_ROUND_UP(ctx->ref_frames[idx].w, 64);
